Got a Block script going and fixed ball sort of
The ball has an impulse affect so it bounces of the stuff better than before however there is a glitch where if it hits 2 objects itll be double flipped and be stuck to the wall.
the blocks have a script now with a hit point meter so it can change materials when we have them. they have 3 hit points gonna need some special affect probably will be off of the at script maybe an inheritance structure